refactoring ui files
